requirements.txt
openai==0.27.2
tiktoken==0.3.3
requests==2.28.2
python-dotenv==1.0.0
